set color_dom
