+++ src/defaultSettings.js
@@ -150,6 +150,7 @@ export function syncSettings(preSettings) {
     nextSettings.legend = nextSettings.legend || { location: 'top' };
     nextSettings.legend.label = nextSettings.color.label;
     nextSettings.legend.order = nextSettings.color.values;
+    nextSettings.legend.color_dom = nextSettings.color.values;
 
     //Default filters
     if (!nextSettings.filters || nextSettings.filters.length === 0) {
